Fix return values of all public functions
All public functions return a negative number in case of error, and 0 or a
positive number in case of success.

Rework DSM error codes, set them more generic.

Add back some assert when it's usefull (You can assert to notify the developer
that he is doing something wrong and not according to the API).

Add smb_session_get_nt_status: get the NT_STATUS returned by the server.

Partially revert 357b7417b23215fde18b6d129f882793b205c69f: smb_tid should stay
a uint16_t, especially if smb_fd is a concatenation of smb_tid and smb_fid.
smb_tid, smb_fd are now returned from arguments in a pointer.

+++ src/smb_trans2.c
@@ -191,7 +191,7 @@ static smb_message  *smb_trans2_find_first (smb_session *s, smb_tid tid, const c
         free(utf_pattern);
         return NULL;
     }
-    msg->packet->header.tid = (uint16_t)tid;
+    msg->packet->header.tid = tid;
 
     SMB_MSG_INIT_PKT(tr2);
     tr2.wct                = 15;
@@ -263,7 +263,7 @@ static smb_message  *smb_trans2_find_next (smb_session *s, smb_tid tid, uint16_t
         free(utf_pattern);
         return NULL;
     }
-    msg_find_next2->packet->header.tid = (uint16_t)tid;
+    msg_find_next2->packet->header.tid = tid;
 
     SMB_MSG_INIT_PKT(tr2_find_next2);
     tr2_find_next2.wct                = 0x0f;
@@ -425,7 +425,7 @@ smb_file  *smb_fstat(smb_session *s, smb_tid tid, const char *path)
         free(utf_path);
         return 0;
     }
-    msg->packet->header.tid = (uint16_t)tid;
+    msg->packet->header.tid = tid;
 
     SMB_MSG_INIT_PKT(tr2);
     tr2.wct                = 15;
